+++ packages/next-adapter/template/babel.config.js
@@ -1,4 +1,3 @@
-// @generated: @expo/next-adapter
 // Learn more: https://github.com/expo/expo/blob/master/docs/pages/versions/unversioned/guides/using-nextjs.md#shared-steps
 
 module.exports = { presets: ['@expo/next-adapter/babel'] };
